+++ monty/operator.py
@@ -1,8 +1,6 @@
 """
 Useful additional functions for operators
 """
-from __future__ import absolute_import, division, unicode_literals
-
 import operator
 
 
@@ -12,8 +10,6 @@ def operator_from_str(op):
 
     raises:
         `KeyError` if invalid string.
-
-    >>> assert operator_from_str("==")(1, 1) and operator_from_str("+")(1,1) == 2
     """
     d = {"==": operator.eq,
          "!=": operator.ne,