+++ tests/test_shutil.py
@@ -1,19 +1,13 @@
-__author__ = 'Shyue Ping Ong'
-__copyright__ = 'Copyright 2014, The Materials Virtual Lab'
-__version__ = '0.1'
-__maintainer__ = 'Shyue Ping Ong'
-__email__ = 'ongsp@ucsd.edu'
-__date__ = '1/24/14'
-
-
 import unittest
 import os
 import shutil
+import platform
+import tempfile
 from gzip import GzipFile
 from io import open
 
 from monty.shutil import copy_r, compress_file, decompress_file, \
-    compress_dir, decompress_dir, gzip_dir
+    compress_dir, decompress_dir, gzip_dir, remove
 
 test_dir = os.path.join(os.path.dirname(__file__), 'test_files')
 
@@ -53,7 +47,6 @@ class CopyRTest(unittest.TestCase):
             txt = f.read()
             self.assertEqual(txt, "what")
 
-
     def tearDown(self):
         shutil.rmtree(os.path.join(test_dir, "cpr_src"))
         shutil.rmtree(os.path.join(test_dir, "cpr_dst"))
@@ -82,6 +75,7 @@ class CompressFileDirTest(unittest.TestCase):
     def tearDown(self):
         os.remove(os.path.join(test_dir, "tempfile"))
 
+
 class GzipDirTest(unittest.TestCase):
 
     def setUp(self):
@@ -100,13 +94,55 @@ class GzipDirTest(unittest.TestCase):
         self.assertFalse((os.path.exists(full_f)))
 
         with GzipFile("{}.gz".format(full_f)) as g:
-            self.assertEquals(g.readline().decode("utf-8"), "what")
+            self.assertEqual(g.readline().decode("utf-8"), "what")
 
         self.assertAlmostEqual(os.path.getmtime("{}.gz".format(full_f)),
                                self.mtime, 4)
 
     def tearDown(self):
         shutil.rmtree(os.path.join(test_dir, "gzip_dir"))
+
+
+class RemoveTest(unittest.TestCase):
+
+    @unittest.skipIf(platform.system() == "Windows", "Skip on windows")
+    def test_remove_file(self):
+        tempdir = tempfile.mkdtemp(dir=test_dir)
+        tempf = tempfile.mkstemp(dir=tempdir)[1]
+        remove(tempf)
+        self.assertFalse(os.path.isfile(tempf))
+        shutil.rmtree(tempdir)
+
+    @unittest.skipIf(platform.system() == "Windows", "Skip on windows")
+    def test_remove_folder(self):
+        tempdir = tempfile.mkdtemp(dir=test_dir)
+        remove(tempdir)
+        self.assertFalse(os.path.isdir(tempdir))
+
+    @unittest.skipIf(platform.system() == "Windows", "Skip on windows")
+    def test_remove_symlink(self):
+        tempdir = tempfile.mkdtemp(dir=test_dir)
+        tempf = tempfile.mkstemp(dir=tempdir)[1]
+
+        os.symlink(tempdir, os.path.join(test_dir, "temp_link"))
+        templink = os.path.join(test_dir, "temp_link")
+        remove(templink)
+        self.assertTrue(os.path.isfile(tempf))
+        self.assertTrue(os.path.isdir(tempdir))
+        self.assertFalse(os.path.islink(templink))
+        remove(tempdir)
+
+    @unittest.skipIf(platform.system() == "Windows", "Skip on windows")
+    def test_remove_symlink_follow(self):
+        tempdir = tempfile.mkdtemp(dir=test_dir)
+        tempf = tempfile.mkstemp(dir=tempdir)[1]
+
+        os.symlink(tempdir, os.path.join(test_dir, "temp_link"))
+        templink = os.path.join(test_dir, "temp_link")
+        remove(templink, follow_symlink=True)
+        self.assertFalse(os.path.isfile(tempf))
+        self.assertFalse(os.path.isdir(tempdir))
+        self.assertFalse(os.path.islink(templink))
 
 
 if __name__ == "__main__":