+++ docs_rst/changelog.rst
@@ -1,6 +1,70 @@
 Change log
 ==========
 
+v3.0.1
+------
+1. Bug fixes for Windows.
+
+v3.0.0
+------
+1. Py3 only version.
+
+v2.0.7
+------
+1. MSONable now supports Enum types. (@mkhorton)
+
+v2.0.6
+------
+1. Revert py27 incompatible fmt spec for loadfn and dumpfn for now. This is
+   a much less common use case.
+
+v2.0.5
+------
+1. Checking for extension type in loadfn and dumpfn now relies on ".json",
+   ".yaml" or ".mpk". Further, now ".yml" and ".yaml" are both recognized as
+   YAML files. (@clegaspi)
+2. A fmt kwarg is now supported in loadfn and dumpfn to specify the format
+   explicitly. (@clegaspi)
+
+v2.0.4
+------
+1. Bug fix for invert MSON caused by `@version`.
+
+v2.0.3
+------
+1. Support for nested MSONAble objects with MontyEncoder and dumpfn.
+   (@davidwaroquiers)
+2. Add @version to MSONAble. (@mkhorton)
+
+v2.0.0
+------
+1. Support for Path object in zopen.
+
+v1.0.5
+------
+1. Bug fix for io.reverse_readfile to ensure txt or binary string.
+
+v1.0.4
+------
+1. monty.shutil.remove which allows symlinks removal. Also improved
+   monty.tempfile.ScratchDir cleanup. (@shyamd)
+
+v1.0.3
+------
+1. Bug fix for reverse_readfile for bz2 files (Alex Urban)
+
+v1.0.2
+------
+1. Misc bug fixes (tempdir on Windows)
+
+v1.0.1
+------
+1. Use CLoader and CDumper by default for speed.
+
+v1.0.0
+------
+1. Ruamel.yaml is now used as the default YAML parser and dumper.
+
 v0.9.8
 ------
 1. Now ScratchDir functions as it should by replacing the original directory.